tests: split the arena and forest region tests by subject
Both files are past the 500-line reading budget, so neither can carry a stamp,
and the next change to either pays to read the whole thing. Each holds several
subjects, which is the usual reason a file gets that long.

`arena/tests.rs` becomes the reference-count funnels, the outgoing edge table,
the deref tag check, and struct-key interning. `regionstore/tests/forest.rs`
becomes adoption, subtree drop, reparenting, and the external-reference rescue —
the shape its sibling themed files already had.

No test changes. Each file opens with the call-out and the governing document
its readers arrive without.
